Split AMD Loader Tests to separate file
